Add separate aux module
